% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.chart.R
\name{bar.chart}
\alias{bar.chart}
\title{Simplified Bar Chart}
\usage{
bar.chart(df, var1, lab = FALSE)
}
\arguments{
\item{df}{data frame to read in.}

\item{var1}{the dependent/outcome variable, \eqn{Y}. The variable of interest that should be plotted.}

\item{lab}{logical (default set to \code{FALSE}). When set to \code{lab = TRUE}, will add frequency label for each bar in chart.}
}
\value{
This function returns the bar chart for \code{var1} in data frame \code{df}.
}
\description{
This function plots a bar chart (bar.chart) on a given data frame.
}
\examples{
data <- mtcars

bar.chart(data,cyl)
}
