% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.calc.R
\name{ci.calc}
\alias{ci.calc}
\title{Simplified Confidence Interval Calculation}
\usage{
ci.calc(df, var1, cl)
}
\arguments{
\item{df}{data frame to read in.}

\item{var1}{the variable of interest for which the CI will be calculated.}

\item{cl}{the desired confidence level (in percentages, ranging from \eqn{1} to \eqn{100}).}
}
\value{
This function returns the mean, lower bound, upper bound, and standard error.
}
\description{
This function calculates the confidence interval (for a given confidence level) for a variable in a given data frame.
}
\examples{
data <- mtcars

ci.calc(data,mpg,95)
}
