% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat}
\alias{cormat}
\title{Simplified Correlation Matrix}
\usage{
cormat(df, formula)
}
\arguments{
\item{df}{data frame to read in.}

\item{formula}{the variables in the regression model, \eqn{Y = X_1 + X_2 + ... + X_m}, written as \code{Y ~ X1 + X2}...}
}
\value{
This function returns a correlation matrix for the variables provided in the formula.
}
\description{
This function creates a correlation (cormat) on a data frame of the variables in an equation.
}
\examples{
data <- mtcars

cormat(data, mpg ~ wt + am)
}
