% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_2014}
\alias{gss_2014}
\title{General Social Survey, 2014}
\format{
A data frame with 2538 observations and 676 variables.
}
\usage{
gss_2014
}
\description{
A data set containing General Social Survey data from 2014
}
\keyword{datasets}
