% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howell_aids_long.R
\docType{data}
\name{howell_aids_long}
\alias{howell_aids_long}
\title{Howell Student AIDS Knowledge Data (Long Form)}
\format{
A data frame with 12 observations and 3 variables.
\tabular{ll}{ \cr
 student \tab student id \cr
 time \tab time point measured \cr
knowledge \tab student AIDS knowledge score (at various time points) \cr
}
}
\usage{
howell_aids_long
}
\description{
This data set, from Howell, measures students' knowledge at three time points, in long form.
}
\keyword{datasets}
