% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howell_aids_wide.R
\docType{data}
\name{howell_aids_wide}
\alias{howell_aids_wide}
\title{Howell Student AIDS Knowledge Data (Wide Form)}
\format{
A data frame with 4 observations and 4 variables.
\tabular{ll}{ \cr
 student \tab student id \cr
 t1 \tab student AIDS knowledge score at time 1 \cr
 t2 \tab student AIDS knowledge score at time 2 \cr
 t3 \tab student AIDS knowledge score at time 3 \cr
}
}
\usage{
howell_aids_wide
}
\description{
This data set, from Howell, measures students' knowledge at three time points, in wide form.
}
\keyword{datasets}
