% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ideology_ERA.R
\docType{data}
\name{Ideology_ERA}
\alias{Ideology_ERA}
\title{Fording's State Ideology Data}
\format{
A data frame with 300 observations and 5 variables.
\tabular{ll}{ \cr
 state \tab state name \cr
 year \tab year measured \cr
 c_ideo \tab citizen ideology score \cr
 s_ideo \tab state level ideology score \cr
 era_status \tab measures the the event: adoption/ratification of the Equal Rights Amendment for a state in a given year. 0 equates to not ratified in state in that year, 1 equates to ratified in state in that year \cr
}
}
\usage{
Ideology_ERA
}
\description{
This data set comes from Richard C. Fording (\url{https://rcfording.com/state-ideology-data/}) and used in Berry et al. (1998). The data set includes state ideology data (measured at the state/legislature and citizen levels), for each year between 1960 and 2018. These data will be updated as Fording updates the data. This data set enables inclusion of a variable measuring state-level policy adoption by year. As an example, the data set also include a variable measuring the ratification of the Equal Rights Amendment as depicted in Soule and King (2006).
}
\keyword{datasets}
