% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.t.R
\name{os.t}
\alias{os.t}
\title{Simplified One Sample T-Test}
\usage{
os.t(df, var1, mu)
}
\arguments{
\item{df}{data frame to read in.}

\item{var1}{the dependent/outcome variable, \eqn{Y}.}

\item{mu}{the population mean, \eqn{\mu}.}
}
\value{
This function returns the summary results table for an one sample t-test, examining the mean differences between \code{var1} (in data frame \code{df}) and the population mean \code{mu}.
}
\description{
This function simplifies the call for the one sample t-test (os.t) on a given data frame.
}
\examples{
data <- mtcars

ttest <- os.t(data,mpg,3)
summary(ttest)

}
