% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.corr.R
\name{p.corr}
\alias{p.corr}
\title{Simplified Correlation}
\usage{
p.corr(df, var1, var2)
}
\arguments{
\item{df}{data frame to read in.}

\item{var1}{the dependent/outcome variable, \eqn{Y}.}

\item{var2}{the main independent/predictor variable, \eqn{X}.}
}
\value{
This function returns the summary results table for a Pearson's correlation, examining the relationship between \code{var1} from data frame \code{df}, and \code{var2}.
}
\description{
This function simplifies the call for Pearson's Product-Moment Correlation Coefficient (p.corr) on a given data frame.
}
\examples{
data <- mtcars

p.corr(data,mpg,wt)
}
