% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.t.R
\name{ps.t}
\alias{ps.t}
\title{Simplified Paired Samples (Repeated Measures) T-Test}
\usage{
ps.t(df, t2, t1, var.equal = TRUE)
}
\arguments{
\item{df}{data frame to read in.}

\item{t2}{the \eqn{t_2} or post-test variable.}

\item{t1}{the \eqn{t_1} or pre-test variable.}

\item{var.equal}{logical (default set to \code{T}). When set to \code{var.equal = F}, will employ Welsh's correction to the t-test (for data that violate the equal variances assumption).}
}
\value{
This function returns the summary results table for an paired samples t-test, examining the mean differences between \code{t2} and \code{t1} (in data frame \code{df}).
}
\description{
This function simplifies the call for the paired samples t-test (ps.t) on a given data frame.
}
\examples{
data <- mtcars
data$mpg2 <- c(21.8,14.7,5.9,24.0,12.2,14.9,33.0,27.7,
29.9,33.9,16.0,14.7,13.3,23.1,38.0,39.7,24.1,33.9,
25.7,24.2,28.3,37.3,20.3,36.0,18.1,32.8,28.7,30.3,6.6,
26.4,35.8,32.8)

ttest <- ps.t(data,mpg2,mpg)
summary(ttest)

}
