% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residplot.R
\name{residplot}
\alias{residplot}
\title{Simplified Residuals Plot}
\usage{
residplot(df, formula)
}
\arguments{
\item{df}{data frame to read in.}

\item{formula}{the variables in the regression model, \eqn{Y = X_1 + X_2 + ... + X_m}, written as \code{Y ~ X1 + X2}...}
}
\description{
This function creates a residual plot (residplot) on a data frame of the variables in an equation.
}
\examples{
data <- mtcars

residplot(data, mpg ~ wt + am)
}
