% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.code.R
\name{revcode}
\alias{revcode}
\title{Reverse Coding for Scales}
\usage{
revcode(df, var, missing = c(""))
}
\arguments{
\item{df}{data frame to read in.}

\item{var}{the variable to be recoded.}

\item{missing}{a list of values in the variable that are ``missing'' values.}
}
\value{
This function updates the data frame with a new variable with the recoded values.
}
\description{
This function applies reverse-coding to a variable of interest.
}
\examples{
data <- GSS2014

revcode(data, amcult)
}
