% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error Calculation}
\usage{
se(var, na.rm = TRUE)
}
\arguments{
\item{var}{variable to read in.}

\item{na.rm}{logical (default set to \code{T}). When set to \code{na.rm = F}, will include NA's in calculation.}
}
\value{
This function returns the standard error for a given variable
}
\description{
This function calculates the standard error for a variable in a given data frame.
}
\examples{
data <- mtcars

se(data$mpg)
}
