% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.chisquare.R
\name{summary.chisquare}
\alias{summary.chisquare}
\title{Summarize Results of chi.sq}
\usage{
\method{summary}{chisquare}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{chi.sq}}.}

\item{...}{Additional parameters to pass on.}
}
\value{
Matrix of values for results from chi square test.
}
\description{
Displays results of chi.sq
}
\examples{
data1 <- mtcars
x2 <- chi.sq(data1, vs, am)

summary(x2)

}
