% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ist.R
\name{summary.ist}
\alias{summary.ist}
\title{Summarize Results of is.t}
\usage{
\method{summary}{ist}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{is.t}}.}

\item{...}{Additional parameters to pass on.}
}
\value{
Matrix of values for results from independent samples t-test.
}
\description{
Displays results of is.t
}
\examples{
data1 <- mtcars
ttest <- is.t(data1, mpg, am)

summary(ttest)

}
