% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ost.R
\name{summary.ost}
\alias{summary.ost}
\title{Summarize Results of os.t}
\usage{
\method{summary}{ost}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{os.t}}.}

\item{...}{Additional parameters to pass on.}
}
\value{
Matrix of values for results from one sample t-test.
}
\description{
Displays results of os.t
}
\examples{
data1 <- mtcars
ttest <- os.t(data1,mpg,3)

summary(ttest)

}
