% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ow.anova.R
\name{summary.ow.anova}
\alias{summary.ow.anova}
\title{Summarize Results of ow.anova}
\usage{
\method{summary}{ow.anova}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{ow.anova}}.}

\item{...}{Additional parameters to pass on.}
}
\value{
Matrix of values for results from One-Way ANOVA test.
}
\description{
Displays results of ow.anova
}
\examples{
data1 <- mtcars
ow <- ow.anova(data1, mpg, cyl)

summary(ow)
}
