% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pst.R
\name{summary.pst}
\alias{summary.pst}
\title{Summarize Results of ps.t}
\usage{
\method{summary}{pst}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{ps.t}}.}

\item{...}{Additional parameters to pass on.}
}
\value{
Matrix of values for results from paired samples t-test.
}
\description{
Displays results of ps.t
}
\examples{
data1 <- mtcars
data1$mpg2 <- c(21.8,14.7,5.9,24.0,12.2,14.9,33.0,27.7,
29.9,33.9,16.0,14.7,13.3,23.1,38.0,39.7,24.1,33.9,
25.7,24.2,28.3,37.3,20.3,36.0,18.1,32.8,28.7,30.3,6.6,
26.4,35.8,32.8)

ttest <- ps.t(data1,mpg2,mpg)

summary(ttest)

}
