% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabs.R
\name{tab}
\alias{tab}
\title{Simplified Crosstabs}
\usage{
tab(df, var1, var2)
}
\arguments{
\item{df}{data frame to read in.}

\item{var1}{a first grouping variable.}

\item{var2}{a second grouping variable.}
}
\value{
This function returns the observed and expected frequencies of a bivariate relationship between \code{var1} and \code{var2} in data frame \code{df}.
}
\description{
This function returns a crosstab (tab) on a given data frame, and using simplified calls within the function for two variables, to return the observed and expected frequencies.
}
\examples{
data <- mtcars

tab(data,mpg,cyl)
}
