% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_Counties.R
\docType{data}
\name{US_Counties}
\alias{US_Counties}
\title{US Counties Information for Merging}
\format{
A data frame with 3104 observations and 8 variables.
\tabular{ll}{ \cr
 countystate \tab proper county name and state name \cr
 state_name \tab proper state name \cr
 county_name \tab proper county name \cr
 county_fips \tab county FIPS \cr
 state_abbv \tab abbreviated state name \cr
state_name_cap \tab capitalized state name \cr
state_name_cap_nominate \tab capitalized state name, shortened (as in DW-NOMINATE data) \cr
state_fips \tab state FIPS \cr
}
}
\usage{
US_Counties
}
\description{
This data set provides common names and abbreviations for U.S. counties to enable merging with various data sets.
}
\keyword{datasets}
