% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_States.R
\docType{data}
\name{US_States}
\alias{US_States}
\title{US States Information for Merging}
\format{
A data frame with 50 observations and 5 variables.
\tabular{ll}{ \cr
 state_name \tab proper state name \cr
 state_abbv \tab abbreviated state name \cr
state_name_cap \tab capitalized state name \cr
state_name_cap_nominate \tab capitalized state name, shortened (as in DW-NOMINATE data) \cr
state_fips \tab state FIPS \cr
}
}
\usage{
US_States
}
\description{
This data set provides common names and abbreviations for U.S. states to enable merging with various data sets.
}
\keyword{datasets}
