% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.calc.R
\name{z.calc}
\alias{z.calc}
\title{Simplified Z Scores}
\usage{
z.calc(df, var1, x)
}
\arguments{
\item{df}{data frame to read in.}

\item{var1}{the variable of interest for which the mean and standard deviations will be calculated.}

\item{x}{the desired raw score to compare with the mean and standard deviation of \code{var1}.}
}
\value{
This function returns the raw score, mean, and z-score for a given raw score.
}
\description{
This function calculates the Z score for a given value, relative to the mean and standard deviation for a variable in a given data frame.
}
\examples{
data <- mtcars

z.calc(data,mpg,12)
}
